Working on a much better sensor fusion solution, since it turns out a basic complementary filter is much more accurate than what I had before (though less stable).
Here we're holding onto that stability by limiting the correction speed to be a factor of the gyro speed (correct while moving), and I'm trusting the accelerometer vector more during slow rotations.

However, I think the main source of error here is due to rounding errors with the quaternions. So, next task is instead of deriving the gravity vector from the orientation quaternion, I'm going to correct the orientation quaternion by our best guess gravity vector.